# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Crawler API
  version: 1.0.0
  description: |
    OpenAPI specification for the Crawler API. Generated from Spring controller mappings
    under `src/main/java/com/crawlerApi/api`. Many endpoints require an authenticated account
    context (JWT) and return domain-specific DTOs for domains, audits, pages, elements, and test records.
servers:
  - url: http://localhost:9080
    description: Local default server port from application.properties
security:
  - bearerAuth: []
...
paths:
...
  /v1/integrations:
    get:
      tags: [Integrations]
      summary: List available integrations
      description: Returns metadata for all available integration types.
      responses:
        '200':
          description: List of integration metadata
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IntegrationMetadata' }
        '401': { description: Authentication required }
  /v1/integrations/{type}:
    get:
      tags: [Integrations]
      summary: Get integration by type
      description: Returns metadata and current account config (masked) for the given integration type.
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string }
          description: Integration type (e.g. jira, slack)
      responses:
        '200':
          description: Integration metadata and config
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationWithConfigResponse' }
        '401': { description: Authentication required }
        '404': { description: Unknown integration type }
  /v1/integrations/{type}/config:
    get:
      tags: [Integrations]
      summary: Get integration config
      description: Returns current account config for the integration type (for backend use).
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Config map
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationConfigMap' }
        '401': { description: Authentication required }
        '404': { description: Unknown integration type or no config }
    put:
      tags: [Integrations]
      summary: Create or update integration config
      description: Validates and saves config for the current account.
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IntegrationConfigMap' }
      responses:
        '200': { description: Config saved }
        '400': { description: Invalid config }
        '401': { description: Authentication required }
        '404': { description: Unknown integration type }
        '405': { description: Slack config is OAuth-only; use /v1/integrations/slack/connect }
    delete:
      tags: [Integrations]
      summary: Delete integration config
      description: Removes stored config for the current account and integration type.
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string }
      responses:
        '200': { description: Config deleted }
        '401': { description: Authentication required }
        '404': { description: Unknown integration type }
  /v1/integrations/{type}/test:
    post:
      tags: [Integrations]
      summary: Test integration connection
      description: Tests connection using the current account's stored config.
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string }
      responses:
        '200': { description: Connection succeeded }
        '502': { description: Connection failed }
        '401': { description: Authentication required }
        '404': { description: Unknown integration type or no config }
  /v1/integrations/slack/connect:
    get:
      tags: [Integrations]
      summary: Start Slack OAuth connect
      description: Returns a Slack authorization URL for user click-to-connect flow.
      responses:
        '200':
          description: Authorization URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorizationUrl: { type: string }
        '401': { description: Authentication required }
        '503': { description: Slack OAuth not configured }
  /v1/integrations/slack/callback:
    get:
      tags: [Integrations]
      summary: Handle Slack OAuth callback
      description: Exchanges Slack OAuth code for access token and stores account integration config.
      parameters:
        - in: query
          name: code
          required: true
          schema: { type: string }
        - in: query
          name: state
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Slack connected
          content:
            application/json:
              schema:
                type: object
                properties:
                  connected: { type: boolean }
                  workspace: { type: string }
        '400': { description: Invalid callback payload/state }
        '401': { description: Authentication required }
        '502': { description: Slack token exchange failed }
        '503': { description: Slack OAuth not configured }
  /v1/integrations/product-board:
    post:
      tags: [Integrations]
      summary: Create Product Board integration token
      description: Create a JWT token for Product Board integration (backward compatibility).
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Account' }
      responses:
        '200': { description: Integration token created, content: { application/json: { schema: { type: object, description: JWT } } } }
        '401': { description: Authentication required }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Account:
      type: object
      description: User account (Auth0-linked)
      properties:
        id: { type: integer, format: int64 }
        userId: { type: string, description: "Auth0 user ID" }
        email: { type: string, format: email }
        name: { type: string }
        apiToken: { type: string }
        onboardedSteps: { type: array, items: { type: string } }
      additionalProperties: true
...
    IntegrationMetadata:
      type: object
      description: Metadata for an integration type
      properties:
        id: { type: string, description: Integration type id (e.g. jira, slack) }
        name: { type: string, description: Display name }
        category: { type: string, description: Category (e.g. Product Management, CI/CD) }
        configSchema:
          type: array
          items: { type: string }
          description: Required config keys for this integration
    IntegrationWithConfigResponse:
      type: object
      description: Integration metadata plus optional masked config for current account
      properties:
        metadata: { $ref: '#/components/schemas/IntegrationMetadata' }
        config:
          type: object
          additionalProperties: true
          description: Masked config (secrets redacted), or null if not configured
    IntegrationConfigMap:
      type: object
      additionalProperties: true
      description: Key-value config for an integration (keys depend on integration type)